+++ Source/PurC/include/purc.h
@@ -66,11 +66,12 @@ typedef struct purc_instance_extra_info {
      * When using a HEADLESS renderer, you should specify a file
      * or a named pipe (FIFO), like `file:///var/tmp/purc-foo-bar-msgs.log`.
      *
-     * When using a PURCTH renderer, you should specify the endpoint name
-     * of the renderer like `//-/<app_name>/<runner_name>`. The endpoint name
-     * will be used to distinguish renderers and interperter instances.
+     * When using a THREAD renderer, you should specify the endpoint name
+     * of the renderer like `edpt://localhost/<app_name>/<runner_name>`.
+     * The endpoint name will be used to communicate between the renderer
+     * and the interperter instances.
      *
-     * When using a PURCMC renderer, you can specify a UNIX domain socket
+     * When using a SOCKET renderer, you can specify a UNIX domain socket
      * or a URI of WebSocket:
      *
      *      - UNIX domain socket: `unix:///var/tmp/xxx.sock`